resume_project/requirements.txt
Django==5.1.1
python-decouple==3.8
huggingface_hub==0.23.2
reportlab==4.2.0
PyMuPDF==1.23.0
python-docx==1.1.0